refactor(cli): move some utilities into a utils folder
