+++ packages/cli/src/commands/remove.ts
@@ -16,7 +16,7 @@ import { exec } from '../exec.js';
 import { version } from '../get-package-info.js';
 import { collectMigrations } from '../collect-migrations.js';
 import { migrationRunner } from '../migration-runner.js';
-import { arrayMapAsync } from '../array-map-async.js';
+import { arrayMapAsync } from '../utils/array-map-async.js';
 import { type GetMigrationsFunction } from '../get-migrations.js';
 import { getStandardReporter } from '../reporters/get.js';
