+++ packages/cli/src/utils/get-duration.ts
