+++ packages/cli/src/utils/sanitize-name.test.ts
