+++ packages/cli/src/utils/sanitize-name.ts
