+++ packages/cli/src/utils/with-leading-period.ts
